+++ include/audio/player/AudioSystem.h
@@ -59,7 +59,7 @@ namespace uaudio
 
 			bool simd = false;
 			double m_DeltaTime = 0.0f;
-			float m_Fps = 0;
+			double m_Fps = 0;
 		private:
 			double m_Time = 0.0;
 			size_t m_Frames = 0;
